requirements.txt
PyQt6
pymodbus
requests
paho-mqtt
pysnmp
pyasn1